pyproject.toml
[project]
name = "graal-posto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "opencv-python",
    "python-dotenv",
    "ultralytics",
]
